etsy-insight-engine/requirements.txt
requests
pandas
numpy
beautifulsoup4
lxml
scikit-learn
openpyxl
python-dotenv
matplotlib
streamlit
